Reorganize types.
