+++ smarts/core/smarts.py
@@ -51,7 +51,7 @@ from .motion_planner_provider import MotionPlannerProvider
 from .provider import Provider, ProviderState
 from .road_map import RoadMap
 from .scenario import Mission, Scenario
-from .sensors import Collision
+from .events import Collision
 from .traffic_history_provider import TrafficHistoryProvider
 from .trajectory_interpolation_provider import TrajectoryInterpolationProvider
 from .trap_manager import TrapManager
@@ -144,7 +144,7 @@ class SMARTS:
         self._vehicle_index = VehicleIndex()
 
         # TODO: Should not be stored in SMARTS
-        self._vehicle_collisions = defaultdict(list)  # list of `Collision` instances
+        self._vehicle_collisions = defaultdict(list)
         self._vehicle_states = []
 
         self._bubble_manager = None
